Revert "fix(ark-runtime): support fractional task duration"
This reverts commit f566e1333e08316f27fe0e31fc1316d7d4cbc320.

+++ volcengine-java-sdk-ark-runtime/src/main/java/com/volcengine/ark/runtime/model/content/generation/GetContentGenerationTaskResponse.java
@@ -72,7 +72,7 @@ public class GetContentGenerationTaskResponse {
     private String outputFormat;
 
     @JsonProperty("duration")
-    private java.lang.Double duration;
+    private java.lang.Long duration;
 
     @JsonProperty("resolution")
     private String resolution;
@@ -254,11 +254,11 @@ public class GetContentGenerationTaskResponse {
         this.outputFormat = outputFormat;
     }
 
-    public java.lang.Double getDuration() {
+    public java.lang.Long getDuration() {
         return duration;
     }
 
-    public void setDuration(java.lang.Double duration) {
+    public void setDuration(java.lang.Long duration) {
         this.duration = duration;
     }
